FIFA World Cup Data Compilation
Scrape historical FIFA World Cup data from inaugural to latest tournaments using
BeautifulSoup. Clean and structure the data, creating a comprehensive and user-friendly
dataset for in-depth analyses. The project aims to provide a self-contained resource for
researchers and enthusiasts interested in exploring the rich history of the FIFA World Cup